fix(brackend) se arregló la lógica por nombre de clases de attention -> waittingqueue
