+++ src/main/java/com/vash/entel/service/WaitingQueueService.java
@@ -1,13 +1,17 @@
 package com.vash.entel.service;
 
+import com.vash.entel.dto.NextPendingTicketResponseDTO;
 import com.vash.entel.model.entity.Ticket_code;
 import com.vash.entel.model.entity.WaitingQueue;
 import com.vash.entel.model.enums.AttentionStatus;
+import org.springframework.http.ResponseEntity;
 
+import java.util.Map;
 import java.util.Optional;
 
 public interface WaitingQueueService {
-    Optional<WaitingQueue> getNextPendingTicket();
-    void updateWaitingQueueStatus(WaitingQueue waitingQueue, AttentionStatus status);
+    Optional<NextPendingTicketResponseDTO> getNextPendingTicket(Integer moduleId);
+    ResponseEntity<Map<String, String>> acceptTicket();
+    ResponseEntity<Map<String, String>> rejectTicket();
     void addTicketToQueue(Ticket_code ticketCode);
 }
